requirements.txt
Pillow>=8.0.0
reportlab>=3.5.0